Duplicate d2tmo device profile into d2ltetmo
Since it also exhibits the same problem as issue 121 (https://github.com/commonsguy/cwac-camera/issues/121) and these devices appear to be identical.

Adding this device profile avoids the image corruption in the demo app.

This seems like a duct tape fix, and it would probably be better to use one of these suggestions:

https://github.com/commonsguy/cwac-camera/issues/175

https://github.com/commonsguy/cwac-camera/issues/201

Another alternative is to use "starts with" matching on device profiles and include a "samsung_d2*" profile, since there are loads more S3 devices which are not currently being matched.